LayoutTests/editing/selection/4983858-expected.txt

| <html>
|   <head>
|   <body>
|     contenteditable="true"
|     "This tests for a bug where selecting a word would select the line break and word before it.  Only the word in the paragraph below should be selected:"
|     "\n"
|     "\n"
|     <div>
|       "<#selection-focus>foo<#selection-anchor>"
|     "\n"
